Remove greenlet for event check
