+++ src/monitoring_service/constants.py
@@ -2,4 +2,4 @@ RATIO_OF_SETTLE_TIMEOUT_BEFORE_MONITOR: float = 0.3
 DEFAULT_REQUIRED_CONFIRMATIONS: int = 10
 MAX_FILTER_INTERVAL: int = 200_000
 DEFAULT_GAS_BUFFER_FACTOR: int = 10
-DEFAULT_GAS_CHECK_WAIT: int = 30
+DEFAULT_GAS_CHECK_BLOCKS: int = 100
